Updated XDR generator to use Long for UInt32

+++ src/main/kotlin/org/tokend/wallet/xdr/XdrTypes.kt
@@ -14777,7 +14777,7 @@ public typealias Uint256 = XdrByteArrayFixed32
 //  typedef unsigned int uint32;
 
 //  ===========================================================================
-public typealias Uint32 = kotlin.Int
+public typealias Uint32 = kotlin.Long
 
 // === xdr source ============================================================
